+++ src/artnetstatic.h
@@ -1,0 +1,69 @@
+#ifdef __ARTNETSTATIC__
+#define __ARTNETSTATIC__
+
+static err_t _udp_connect_api(struct tcpip_api_call_data *api_call_msg)
+{
+    udp_api_call_t *msg = (udp_api_call_t *)api_call_msg;
+    msg->err = udp_connect(msg->pcb, msg->addr, msg->port);
+    return msg->err;
+}
+
+static err_t _udp_connect(struct udp_pcb *pcb, const ip_addr_t *addr, u16_t port)
+{
+    udp_api_call_t msg;
+    msg.pcb = pcb;
+    msg.addr = addr;
+    msg.port = port;
+    tcpip_api_call(_udp_connect_api, (struct tcpip_api_call_data *)&msg);
+    return msg.err;
+}
+
+static err_t _udp_disconnect_api(struct tcpip_api_call_data *api_call_msg)
+{
+    udp_api_call_t *msg = (udp_api_call_t *)api_call_msg;
+    msg->err = 0;
+    udp_disconnect(msg->pcb);
+    return msg->err;
+}
+
+static void _udp_disconnect(struct udp_pcb *pcb)
+{
+    udp_api_call_t msg;
+    msg.pcb = pcb;
+    tcpip_api_call(_udp_disconnect_api, (struct tcpip_api_call_data *)&msg);
+}
+
+static err_t _udp_remove_api(struct tcpip_api_call_data *api_call_msg)
+{
+    udp_api_call_t *msg = (udp_api_call_t *)api_call_msg;
+    msg->err = 0;
+    udp_remove(msg->pcb);
+    return msg->err;
+}
+
+static void _udp_remove(struct udp_pcb *pcb)
+{
+    udp_api_call_t msg;
+    msg.pcb = pcb;
+    tcpip_api_call(_udp_remove_api, (struct tcpip_api_call_data *)&msg);
+}
+
+static err_t _udp_bind_api(struct tcpip_api_call_data *api_call_msg)
+{
+    udp_api_call_t *msg = (udp_api_call_t *)api_call_msg;
+    msg->err = udp_bind(msg->pcb, msg->addr, msg->port);
+    return msg->err;
+}
+
+static err_t _udp_bind(struct udp_pcb *pcb, const ip_addr_t *addr, u16_t port)
+{
+    udp_api_call_t msg;
+    msg.pcb = pcb;
+    msg.addr = addr;
+    msg.port = port;
+    tcpip_api_call(_udp_bind_api, (struct tcpip_api_call_data *)&msg);
+    return msg.err;
+}
+
+
+#endif
